+++ crates/pvm/parser/src/format/ri.rs
@@ -21,7 +21,7 @@ impl From<&[u8]> for RI {
 
         RI {
             reg0: (bytes[0] % 16).min(12),
-            imm0: u32::read_imm(&bytes[1..]),
+            imm0: u64::read_imm(&bytes[1..]),
         }
     }
 }
